Code organisation improvement
