+++ Extensions/UIImageViewExtension.swift
@@ -1,0 +1,25 @@
+//
+//  UIImageViewExtension.swift
+//  moviesApp
+//
+//  Created by MBP on 19/03/2021.
+//
+
+import Foundation
+import SwiftUI
+
+extension UIImageView {
+    
+    func load(url: String, baseUrl: String ) {
+        let url = URL(string: baseUrl + url)!
+        DispatchQueue.global().async { [weak self] in
+            if let data = try? Data(contentsOf: url) {
+                if let image = UIImage(data: data) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            }
+        }
+    }
+}
