+++ moviesApp/MoviesCollectionViewCell.swift
@@ -10,24 +10,10 @@ import UIKit
 class MoviesCollectionViewCell: UICollectionViewCell {
     @IBOutlet var label: UILabel!
     @IBOutlet var image: UIImageView!
- 
+    
     func setup(with result: Result) {
-        image.load(url: result.posterPath)
+        image.load(url: result.posterPath, baseUrl: Constants.Design.Image.urlBase)
         label.text = result.title
     }
 }
-extension UIImageView {
-    func load(url: String) {
 
-        let url = URL(string: Constants.Design.Image.urlBase + url)!
-        DispatchQueue.global().async { [weak self] in
-            if let data = try? Data(contentsOf: url) {
-                if let image = UIImage(data: data) {
-                    DispatchQueue.main.async {
-                        self?.image = image
-                    }
-                }
-            }
-        }
-    }
-}
